Moved MapReduce-specific utilities into io.bespin.scala.mapreduce.util
