+++ src/main/scala/io/bespin/scala/mapreduce/util/MapReduceSugar.scala
@@ -1,4 +1,4 @@
-package io.bespin.scala.mapreduce.bigram
+package io.bespin.scala.mapreduce.util
 
 import io.bespin.scala.util.WritableConversions
 import org.apache.hadoop.conf.Configuration
@@ -37,6 +37,7 @@ trait WrappedWithClazz[T] {
 /**
   * A wrapped instance of a MapReduce Mapper object which encapsulates the type arguments of the Mapper
   * with TypeTags and provides a relatively type-safe way to get the class of the Mapper
+  *
   * @tparam KI Key Input type
   * @tparam VI Value Input type
   * @tparam KO Key Output type
@@ -49,6 +50,7 @@ case class WrappedMapper[KI:TT,VI:TT,KO:TT,VO:TT](m: Mapper[KI,VI,KO,VO]) extend
 /**
   * A wrapped instance of a MapReduce Reducer object which encapsulates the type arguments of the Reducer
   * with TypeTags and provides a relatively type-safe way to get the class of the Reducer
+  *
   * @tparam KI Key Input type
   * @tparam VI Value Input type
   * @tparam KO Key Output type
@@ -61,6 +63,7 @@ case class WrappedReducer[KI:TT,VI:TT,KO:TT,VO:TT](r: Reducer[KI,VI,KO,VO]) exte
 /**
   * A wrapped instance of a MapReduce Partitioner object which encapsulates the type arguments of the Partitioner
   * with TypeTags and provides a relatively type-safe way to get the class of the Partitioner
+  *
   * @tparam KI Key Input type
   * @tparam VI Value Input type
   */
@@ -79,6 +82,7 @@ sealed trait HasJob {
 /**
   * BaseJob is the root class for creating a MapReduce job using the syntactic sugar provided in this
   * and related classes.
+  *
   * @param name Name of the Hadoop job.
   * @param initialConfig Optional initial Configuration (can be None)
   * @param callingClass Class symbol of the calling class (needed by Hadoop to determine the source jar)
@@ -94,6 +98,7 @@ case class BaseJob(name: String, initialConfig: Option[Configuration] = None, ca
 
 /**
   * An InputDefinition describes the input format of a particular Job
+  *
   * @tparam KI Key Input type
   * @tparam VI Value Input type
   */
@@ -103,12 +108,14 @@ sealed trait InputDefinition[KI,VI] extends HasJob
   * An OutputDefinition describes the output format of a particular Job.
   * Once an OutputDefinition (and all its prerequisites) have been fully defined, the Job can be
   * run as a MapReduce job.
+  *
   * @tparam KO Key Output type
   * @tparam VO Value Output type
   */
 sealed trait OutputDefinition[KO,VO] extends HasJob {
   /**
     * Runs the constructed MapReduce Job
+    *
     * @param verbose Enables verbose logging output
     * @param mirror Implicit type mirror used for resolving runtime types of arguments
     */
@@ -150,6 +157,7 @@ case class FileOutputDefinition[KO:TT,VO:TT](reduceStage: ReduceStage[_,_,KO,VO]
 /**
   * A class representing the "map" stage of a MapReduce job.
   * Also contains optional Partitioner and Combiner objects.
+  *
   * @param inputDefinition Reference to the "input" stage of the job, which must be defined before the mapper
   * @param map Wrapped mapper object
   * @param combine Optional wrapped combiner(reducer) object
@@ -167,6 +175,7 @@ case class MapStage[KI:TT,VI:TT,KO:TT,VO:TT](inputDefinition: InputDefinition[KI
 
   /**
     * Creates a copy of this stage with the combiner set
+    *
     * @param c Wrapped combiner (reducer) object
     */
   def combine(c: WrappedReducer[KO,VO,KO,VO]): MapStage[KI,VI,KO,VO] =
@@ -174,6 +183,7 @@ case class MapStage[KI:TT,VI:TT,KO:TT,VO:TT](inputDefinition: InputDefinition[KI
 
   /**
     * Creates a copy of this stage with the partitioning behaviour set
+    *
     * @param p Wrapped partitioner object
     */
   def partitionBy(p: WrappedPartitioner[KO,VO]): MapStage[KI,VI,KO,VO] =
@@ -181,6 +191,7 @@ case class MapStage[KI:TT,VI:TT,KO:TT,VO:TT](inputDefinition: InputDefinition[KI
 
   /**
     * Creates a "reduce" stage with this map stage as the dependency
+    *
     * @param reducer Wrapped reducer object
     * @param numReduceTasks Number of reduce tasks (default = 1)
     * @tparam RKO Reducer Key Output
@@ -202,6 +213,7 @@ case class MapStage[KI:TT,VI:TT,KO:TT,VO:TT](inputDefinition: InputDefinition[KI
 
 /**
   * A class representing the "reduce" stage of a MapReduce job
+  *
   * @param mapStage Reference to the "map" stage of the job, which must be defined before the reduce
   * @param reduce Wrapped reducer object
   * @param numReduceTasks Number of reducers (default 1)
@@ -217,6 +229,7 @@ case class ReduceStage[KI:TT,VI:TT,KO:TT,VO:TT](mapStage: MapStage[_,_,KI,VI],
 
   /**
     * Specify the combiner behaviour (this creates a new copy of the "map" dependency)
+    *
     * @param c Wrapped reducer object to use as combiner
     */
   def combine(c: WrappedReducer[KI,VI,KI,VI]): ReduceStage[KI,VI,KO,VO] =
@@ -224,6 +237,7 @@ case class ReduceStage[KI:TT,VI:TT,KO:TT,VO:TT](mapStage: MapStage[_,_,KI,VI],
 
   /**
     * Specify the partitioning behaviour (this creates a new copy of the "map" dependency)
+    *
     * @param p Wrapped partitioner object to use as partitioner
     */
   def partitionBy(p: WrappedPartitioner[KI,VI]): ReduceStage[KI,VI,KO,VO] =
@@ -270,6 +284,7 @@ trait StageSyntax extends WrappingSyntax {
 
   /**
     * Creates a BaseJob instance which can be chained with other operations to configure a MapReduce Job.
+    *
     * @param name Name of the MapReduce job
     * @param caller Class of the invoking class. This can be provided implicitly, such as in the case where
     *               the enclosing class extends [[WithCallingClass]]
@@ -278,6 +293,7 @@ trait StageSyntax extends WrappingSyntax {
 
   /**
     * Creates a BaseJob instance which can be chained with other operations to configure a MapReduce Job.
+    *
     * @param name Name of the MapReduce job
     * @param initialConfig Configuration object to supply to the constructor of the contained MapReduce Job
     * @param caller Class of the invoking class. This can be provided implicitly, such as in the case where
@@ -290,6 +306,7 @@ trait StageSyntax extends WrappingSyntax {
       * file is a general-purpose means of reading an input file, but requires the specification of a
       * FileInputFormat of the correct type in order to be applied. The type of the file determines what
       * the type constraints on the input types of the Mapper will be.
+      *
       * @param path Path of the file to read in
       * @param inputFormat Hadoop FileInputFormat[KI,VI] which determines how to parse the file
       * @tparam KI Key Input type of the file - Will determine the input key type of the Mapper stage
@@ -303,6 +320,7 @@ trait StageSyntax extends WrappingSyntax {
     /**
       * textFile creates a FileInputDefinition specifically using the TextInputFormat. The resulting
       * key type is LongWritable, and the value type is Text.
+      *
       * @param path Path of the file to read in
       */
     def textFile(path: Path): FileInputDefinition[LongWritable, Text] = {
@@ -313,6 +331,7 @@ trait StageSyntax extends WrappingSyntax {
   implicit class InputStageSyntax[KI:TT,VI:TT](inputStage: InputDefinition[KI,VI]) {
     /**
       * Create a MapStage from an input stage
+      *
       * @param map Wrapped Mapper object to set as the Mapper for this job
       * @tparam KO Key Output type of the mapper
       * @tparam VO Value Output type of the mapper
@@ -324,6 +343,7 @@ trait StageSyntax extends WrappingSyntax {
     /**
       * Bind the output path of the job to a file, and optionally delete the file(s) currently present
       * at that location
+      *
       * @param path Path of output directory
       * @param deleteExisting If true, deletes the contents of the output directory before running the job
       */
